extract the menu bar into a component
