+++ src/app/components/mod.rs
@@ -1,0 +1,7 @@
+pub mod menu_bar;
+
+pub trait AppComponent {
+    type Context;
+
+    fn add(ctx: &mut Self::Context, ui: &mut eframe::egui::Ui);
+}
